// src/app/core/api/generated/schemas/couponsControllerListParams.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * stallio-auto-parts merged API
 * OpenAPI spec version: 1.0
 */
import type { CouponsControllerListAppliesTo } from "./couponsControllerListAppliesTo";
import type { CouponsControllerListCouponType } from "./couponsControllerListCouponType";
import type { CouponsControllerListCustomerEligibility } from "./couponsControllerListCustomerEligibility";
import type { CouponsControllerListStatus } from "./couponsControllerListStatus";

export type CouponsControllerListParams = {
  page?: number;
  limit?: number;
  search?: string;
  coupon_type?: CouponsControllerListCouponType;
  customer_eligibility?: CouponsControllerListCustomerEligibility;
  applies_to?: CouponsControllerListAppliesTo;
  status?: CouponsControllerListStatus;
  stackable?: boolean;
  free_shipping?: boolean;
  auto_apply?: boolean;
  is_valid?: boolean;
  coupon_value_min?: number;
  coupon_value_max?: number;
  min_order_amount?: number;
  max_discount_amount?: number;
  usage_limit?: number;
  per_customer_limit?: number;
  used_count_min?: number;
  used_count_max?: number;
  category_id?: number;
  product_id?: number;
  dateFrom?: string;
  dateTo?: string;
  startDateFrom?: string;
  startDateTo?: string;
  expireDateFrom?: string;
  expireDateTo?: string;
};