// src/app/core/api/generated/schemas/paginatedCouponsResponseDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * stallio-auto-parts merged API
 * OpenAPI spec version: 1.0
 */
import type { CouponsResponseDto } from "./couponsResponseDto";
import type { PaginatedCouponsResponseDtoMeta } from "./paginatedCouponsResponseDtoMeta";

export interface PaginatedCouponsResponseDto {
  data: CouponsResponseDto[];
  meta: PaginatedCouponsResponseDtoMeta;
}