// src/app/core/api/generated/schemas/paginatedProductCategoryResponseDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * stallio-auto-parts merged API
 * OpenAPI spec version: 1.0
 */
import type { PaginatedProductCategoryResponseDtoMeta } from "./paginatedProductCategoryResponseDtoMeta";
import type { ProductCategoryResponseDto } from "./productCategoryResponseDto";

export interface PaginatedProductCategoryResponseDto {
  data: ProductCategoryResponseDto[];
  meta: PaginatedProductCategoryResponseDtoMeta;
}