// src/app/core/api/generated/schemas/productSearchSuggestionDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * stallio-auto-parts merged API
 * OpenAPI spec version: 1.0
 */
import type { ProductSearchSuggestionDtoMeta } from "./productSearchSuggestionDtoMeta";
import type { ProductSearchSuggestionDtoType } from "./productSearchSuggestionDtoType";

export interface ProductSearchSuggestionDto {
  type: ProductSearchSuggestionDtoType;
  id: number;
  label: string;
  meta?: ProductSearchSuggestionDtoMeta;
}